ESD-4067-no-hairlines-on-edge-aligned-paths
The svg contrib device truncated the page's MediaSize to an integer number of
points when writing the SVG viewBox and page clip rect, while the declared
physical width/height kept full floating-point precision. Any fractional
remainder in the page size (common for custom page sizes converted from
inches/mm) desynced the two, which distorted the SVG-units-per-inch ratio
downstream consumers derive from viewBox vs. physical size.

- svg_write_header: emit viewBox using the untruncated MediaSize (%lf instead
  of (int) truncation)
- svg_beginpage: apply the same fix to the page clip rect's width/height

+++ contrib/gdevsvg.c
@@ -1227,9 +1227,9 @@ svg_write_header(gx_device_svg* svg)
 		SVG_XMLNS, SVG_VERSION);
 	/* svg_write(svg, line); */
 	svg_write_bytes_sputs(svg, line, strlen(line));
-	gs_snprintf(line, sizeof(line), "\n\twidth='%.3fin' height='%.3fin' viewBox='0 0 %d %d'>\n",
+	gs_snprintf(line, sizeof(line), "\n\twidth='%.3fin' height='%.3fin' viewBox='0 0 %lf %lf'>\n",
 		(double)svg->MediaSize[0] / 72.0, (double)svg->MediaSize[1] / 72.0,
-		(int)svg->MediaSize[0], (int)svg->MediaSize[1]);
+		(double)svg->MediaSize[0], (double)svg->MediaSize[1]);
 	svg_write_bytes_sputs(svg, line, strlen(line));
 
 	/*
@@ -1451,8 +1451,8 @@ svg_beginpage(gx_device_vector* vdev)
 	// Write clip rect for the new page
 	gs_snprintf(line, sizeof(line), "<clipPath id='clip%i'>\n", ++svg->highestUsedId);
 	svg_write_bytes_sputs(svg, line, strlen(line));
-	gs_snprintf(line, sizeof(line), "<rect x='%d' y='%d' width='%d' height='%d' stroke='none' fill='none'/>\n",
-		0, 0, (int)svg->MediaSize[0], (int)svg->MediaSize[1]);
+	gs_snprintf(line, sizeof(line), "<rect x='%d' y='%d' width='%lf' height='%lf' stroke='none' fill='none'/>\n",
+		0, 0, (double)svg->MediaSize[0], (double)svg->MediaSize[1]);
 	svg_write_bytes_sputs(svg, line, strlen(line));
 	svg_write_sputs(svg, "</clipPath>\n");
